MenuDisplay/backend/requirements.txt
paho-mqtt
python-dotenv
requests
bs4